Push shared files down from feature modules
