+++ enterprise/frontend/src/metabase-enterprise/caching/constants.ts
@@ -9,7 +9,7 @@ import {
   PerformanceTabId,
   type StrategyData,
 } from "metabase/admin/performance/types";
-import { defaultCron } from "metabase/admin/performance/utils";
+import { defaultCron } from "metabase/common/components/Schedule/cron";
 import { CacheDurationUnit } from "metabase-types/api";
 import type { AdminPath } from "metabase-types/store";
