+++ frontend/src/metabase/common/components/QuestionDownloadWidget/index.ts
@@ -1,0 +1,4 @@
+export { QuestionDownloadWidget } from "./QuestionDownloadWidget";
+export type { FormatPreference } from "./QuestionDownloadWidget";
+export { useDownloadData } from "./use-download-data";
+export type { UseDownloadDataParams } from "./use-download-data";
