+++ frontend/src/metabase/common/components/Schedule/Schedule.tsx
@@ -4,10 +4,6 @@ import { match } from "ts-pattern";
 import { c, msgid, ngettext } from "ttag";
 import _ from "underscore";
 
-import {
-  cronToScheduleSettings,
-  scheduleSettingsToCron,
-} from "metabase/admin/performance/utils";
 import { CronExpressionInput } from "metabase/common/components/CronExpressioInput";
 import { formatCronExpressionForUI } from "metabase/lib/cron";
 import { removeNullAndUndefinedValues } from "metabase/lib/types";
@@ -27,6 +23,7 @@ import {
   SelectWeekday,
   SelectWeekdayOfMonth,
 } from "./components";
+import { cronToScheduleSettings, scheduleSettingsToCron } from "./cron";
 import { byTheMinuteIntervals } from "./strings";
 import type { UpdateSchedule } from "./types";
 import { getScheduleDefaults } from "./utils";
