+++ frontend/src/metabase/common/components/Sidebar/Sidebar.module.css
