+++ frontend/src/metabase/common/components/Sidebar/index.ts
@@ -1,0 +1,1 @@
+export { Sidebar, SIDEBAR_WIDTH } from "./Sidebar";
