+++ frontend/src/metabase/dashboard/components/DashCard/DashCardMenu/utils.ts
@@ -1,7 +1,5 @@
-import { PLUGIN_FEATURE_LEVEL_PERMISSIONS } from "metabase/plugins";
 import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
-import type { Dataset } from "metabase-types/api";
 
 export const canEditQuestion = (question: Question) => {
   // question.query() is an expensive call, short circuit when possible
@@ -11,12 +9,4 @@ export const canEditQuestion = (question: Question) => {
 
   const { isEditable } = Lib.queryDisplayInfo(question.query());
   return isEditable;
-};
-
-export const canDownloadResults = (result?: Dataset) => {
-  return (
-    !!result?.data &&
-    !result.error &&
-    PLUGIN_FEATURE_LEVEL_PERMISSIONS.canDownloadResults(result)
-  );
 };
