+++ frontend/src/metabase/dashboard/utils.ts
@@ -1,11 +1,10 @@
 import type { Location } from "history";
 import _ from "underscore";
 
+import { isQuestionDashCard, isVirtualDashCard } from "metabase/lib/dashboard";
 import { SERVER_ERROR_TYPES } from "metabase/lib/errors";
 import { isStaticEmbeddingEntityLoadingError } from "metabase/lib/errors/is-static-embedding-entity-loading-error";
 import type { StaticEmbeddingEntityError } from "metabase/lib/errors/types";
-import { isJWT } from "metabase/lib/utils";
-import { isUuid } from "metabase/lib/uuid";
 import {
   getGenericErrorMessage,
   getPermissionErrorMessage,
@@ -87,35 +86,14 @@ export function expandInlineCard(card?: Card | VirtualCard) {
   };
 }
 
-export function isQuestionCard(card: Card | VirtualCard) {
-  // Some old virtual cards have dataset_query equal to {} so we need to check for null and empty object
-  return (
-    card.dataset_query != null && Object.keys(card.dataset_query).length > 0
-  );
-}
-
-export function isQuestionDashCard(
-  dashcard: BaseDashboardCard,
-): dashcard is QuestionDashboardCard {
-  return (
-    "card_id" in dashcard &&
-    "card" in dashcard &&
-    !isVirtualDashCard(dashcard) &&
-    !isActionDashCard(dashcard)
-  );
-}
-
-export function isActionDashCard(
-  dashcard: BaseDashboardCard,
-): dashcard is ActionDashboardCard {
-  return "action" in dashcard;
-}
-
-export function isVirtualDashCard(
-  dashcard: Pick<BaseDashboardCard, "visualization_settings">,
-): dashcard is VirtualDashboardCard {
-  return _.isObject(dashcard?.visualization_settings?.virtual_card);
-}
+export {
+  getDashboardType,
+  isActionDashCard,
+  isQuestionCard,
+  isQuestionDashCard,
+  isTransientId,
+  isVirtualDashCard,
+} from "metabase/lib/dashboard";
 
 export function getVirtualCardType(dashcard: BaseDashboardCard) {
   return dashcard?.visualization_settings?.virtual_card?.display;
@@ -241,25 +219,6 @@ export function getCurrentTabDashboardCards(
 
 export function hasDatabaseActionsEnabled(database: Database) {
   return database.settings?.["database-enable-actions"] ?? false;
-}
-
-export function isTransientId(id: unknown) {
-  return typeof id === "string" && /\/auto\/dashboard/.test(id);
-}
-
-export function getDashboardType(id: unknown) {
-  if (id == null || typeof id === "object") {
-    // HACK: support inline dashboards
-    return "inline";
-  } else if (isUuid(id)) {
-    return "public";
-  } else if (isJWT(id)) {
-    return "embed";
-  } else if (isTransientId(id)) {
-    return "transient";
-  } else {
-    return "normal";
-  }
 }
 
 export async function fetchDataOrError<T>(dataPromise: Promise<T>) {
