+++ frontend/src/metabase/notifications/NewPulseSidebar.tsx
@@ -1,7 +1,7 @@
 import { t } from "ttag";
 
 import { Link } from "metabase/common/components/Link";
-import { Sidebar } from "metabase/dashboard/components/Sidebar";
+import { Sidebar } from "metabase/common/components/Sidebar";
 import { useSelector } from "metabase/lib/redux";
 import { ChannelCard } from "metabase/notifications/pulse/components/ChannelCard";
 import { getApplicationName } from "metabase/selectors/whitelabel";
