+++ frontend/src/metabase/parameters/utils/mapping-options.ts
@@ -7,7 +7,7 @@ import {
   isActionDashCard,
   isQuestionDashCard,
   isVirtualDashCard,
-} from "metabase/dashboard/utils";
+} from "metabase/lib/dashboard";
 import { getGroupName } from "metabase/querying/filters/utils/groups";
 import { getAllowedIframeAttributes } from "metabase/visualizations/visualizations/IFrameViz/utils";
 import * as Lib from "metabase-lib";
