+++ frontend/src/metabase/query_builder/components/QuestionDownloadPopover/QuestionDownloadPopover.tsx
@@ -2,10 +2,6 @@ import cx from "classnames";
 import { useState } from "react";
 import { t } from "ttag";
 
-import type {
-  ExportFormat,
-  TableExportFormat,
-} from "metabase/common/types/export";
 import { useEmbeddingEntityContext } from "metabase/embedding/context";
 import { PLUGIN_FEATURE_LEVEL_PERMISSIONS } from "metabase/plugins";
 import {
@@ -18,13 +14,16 @@ import {
 } from "metabase/ui";
 import type { Dataset } from "metabase-types/api";
 
-import { QuestionDownloadWidget } from "../QuestionDownloadWidget";
 import {
+  type FormatPreference,
+  QuestionDownloadWidget,
   type UseDownloadDataParams,
   useDownloadData,
-} from "../QuestionDownloadWidget/use-download-data";
+} from "../QuestionDownloadWidget";
 
 import S from "./QuestionDownloadPopover.module.css";
+
+export type { FormatPreference };
 
 export type QuestionDownloadPopoverProps = {
   className?: string;
@@ -32,11 +31,6 @@ export type QuestionDownloadPopoverProps = {
 } & Pick<UseDownloadDataParams, "question" | "result"> &
   Pick<ActionIconProps, "variant"> &
   Partial<Omit<UseDownloadDataParams, "question" | "result">>;
-
-export type FormatPreference = {
-  last_download_format: ExportFormat;
-  last_table_download_format: TableExportFormat;
-};
 
 export type BaseQuestionDownloadPopoverProps = QuestionDownloadPopoverProps & {
   formatPreference?: FormatPreference;
