+++ frontend/src/metabase/query_builder/components/QuestionDownloadWidget/index.ts
@@ -1,1 +1,6 @@
-export { QuestionDownloadWidget } from "./QuestionDownloadWidget";
+export {
+  QuestionDownloadWidget,
+  type FormatPreference,
+  useDownloadData,
+  type UseDownloadDataParams,
+} from "metabase/common/components/QuestionDownloadWidget";
