+++ frontend/src/metabase/visualizations/lib/settings/visualization.ts
@@ -1,7 +1,7 @@
 import { assocIn } from "icepick";
 import { t } from "ttag";
 
-import { isVirtualDashCard } from "metabase/dashboard/utils";
+import { isVirtualDashCard } from "metabase/lib/dashboard";
 import { getVisualizationRaw } from "metabase/visualizations";
 import { trackCardSetToHideWhenNoResults } from "metabase/visualizations/lib/settings/analytics";
 import type {
